+++ flatland/envs/schedule_utils.py
@@ -1,8 +1,10 @@
-import collections
+from typing import List, NamedTuple
 
-Schedule = collections.namedtuple('Schedule',   'agent_positions '
-                                                'agent_directions '
-                                                'agent_targets '
-                                                'agent_speeds '
-                                                'agent_malfunction_rates '
-                                                'max_episode_steps')
+from flatland.core.grid.grid4 import Grid4TransitionsEnum
+from flatland.core.grid.grid_utils import IntVector2DArray
+
+Schedule = NamedTuple('Schedule', [('agent_positions', IntVector2DArray),
+                                   ('agent_directions', List[Grid4TransitionsEnum]),
+                                   ('agent_targets', IntVector2DArray),
+                                   ('agent_speeds', List[float]),
+                                   ('agent_malfunction_rates', List[int])])